requirements.txt
fastapi
fastapi[standard]
groq
uvicorn
fastapi-cors
redis[hiredis]
redis
python-dotenv